refactor(emitter): split private-field new helpers
